llm-analyzer/src/requirements.txt
beautifulsoup4==4.13.4
langchain==0.3.24
langchain_community==0.3.22
langchain_ollama==0.3.2
pdfminer.six==20231228
pypdf==5.4.0
Requests==2.32.3
toml==0.10.2
unstructured==0.17.0
urllib3==2.4.0
